+++ resources/views/layouts/admin.blade.php
@@ -1,1 +1,0 @@
-@include('webblocks-cms::layouts.admin', get_defined_vars())